Breaking up the Builder
